reporte identificación de usuario
