+++ src/app/models/report/report-director.model.ts
@@ -1,5 +1,6 @@
 import { StepBuilder } from "../steps/step-builder.model";
-import { Report } from "./report.model";
+import { ReportHeaderBuilder } from "./report-header.model";
+import { Report, ReportPageA4, ReportTable } from "./report.model";
 
 export class ReportEssayDirector {
   private steps!: StepBuilder[];
@@ -11,11 +12,17 @@ export class ReportEssayDirector {
   public createReport(): Report {
     const report: Report = new Report();
 
+    const pageA4 = new ReportPageA4();
+    const reportHeaderBuilder = new ReportHeaderBuilder();
+    const header = reportHeaderBuilder.produce();
+    pageA4.add(header);
+
     this.steps.forEach(({ reportBuilder }) => {
-      const stepReport = reportBuilder.produce();
-      report.pages = report.pages.concat(stepReport.pages);
+      const stepReportTable: ReportTable = reportBuilder.produce();
+      pageA4.add(stepReportTable);
     });
 
+    report.pages.push(pageA4);
     return report;
   }
 }
