+++ src/app/models/steps/preparation.step.model.ts
@@ -2,7 +2,6 @@ import { ReplaySubject } from "rxjs";
 import { EssayTemplateStep } from "../database/tables/essay-template-step.model";
 import { Action } from "../actions/action.model";
 import { StepBuilder } from "./step-builder.model";
-import { UserIdentificationAction } from "../actions/user-identification.action.model";
 import { StandIdentificationAction } from "../actions/stand-identification.action.model";
 import { StandIdentificationMinimalFieldsDecorator } from "../actions/meter-identification.action.model";
 import { ReportBuilder } from "../report/report-builder.model";
@@ -11,12 +10,10 @@ export class PreparationStep extends StepBuilder {
 
   constructor(essayTemplateStep: EssayTemplateStep, destroyed$: ReplaySubject<boolean>) {
     const reportBuilder = new ReportBuilder();
-    const userIdentificationAction = new UserIdentificationAction();
     const standIdentificationAction = new StandIdentificationAction(destroyed$);
     const standIdentificationMinimalFieldsDecorator = new StandIdentificationMinimalFieldsDecorator(standIdentificationAction);
 
     const _actions: Action[] = [
-      userIdentificationAction,
       standIdentificationAction,
     ];
